requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.36
pymysql==1.1.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.5.2
python-multipart==0.0.12
email-validator==2.2.0
pytest==8.3.3
httpx==0.27.2
jinja2==3.1.4
pillow==11.0.0
